Fixes after feedback
