+++ .storybook/src/features/index.stories.ts
@@ -8,7 +8,6 @@ export * from './wellborepath-layer.stories';
 export * from './casing-layer.stories';
 export * from './cement-layer.stories';
 export * from './completion-layer.stories';
-export * from './completion-layer-v2.stories';
 export * from './holesize-layer.stories';
 export * from './geomodel-layer.stories';
 export * from './casing-and-cement.stories';
