+++ src/layers/index.ts
@@ -14,6 +14,5 @@ export * from './CasingLayer';
 export * from './CementLayer';
 export * from './SeismicCanvasLayer';
 export * from './base/PixiLayer';
-export * from './CompletionLayerV2';
 export * from './CompletionLayer';
 export * from './CasingAndCementLayer';
